requirements.txt
backcall==0.1.0
decorator==4.2.1
Django==1.10
django-filter==1.1.0
django-markdown-deux==1.0.5
django-pagedown==1.0.4
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
ipdb==0.11
ipython==5.6.0
ipython-genutils==0.2.0
jedi==0.11.1
Markdown==2.6.11
markdown2==2.3.5
parso==0.1.1
pexpect==4.4.0
pickleshare==0.7.4
Pillow==5.0.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
wcwidth==0.1.7
